- All functionality in SimpleMapper is now re-implemented in a clean way SimpleMapper will be removed after the new code is tested.
- The access control is now cleaner and closer to the spec. We do check
the method, and it's easy to extend. I think it's too early to try to
optimize this code, probably it's more important to integrate with apache
and other web servers.

- Start working on SecurityCheck - it will be replaced with independent
modules, better design, etc.

- added the request notes to the interface. The notes are faster attributes
( indexed array instead of hashtable), and we need to keep internal attribs
separate from servlet-visible attributes.