Improved security ( first part ).
- extracting the credentials from requests moved to SecurityTools ( should go to
a "callback" system like in JAAS )

- Authorization/ACL is separated from the details of user interaction

- User interaction is simple servlets, one per method ( Basic, Form). Form is not
tested ( probably doesn't work ).

- a bit of cleanup in error handling - clear signature for exception/status code
errors, etc. Also, add body to 404.

- Removed RequestSecurityProvider:
	- user Principal can be set by the security interceptor, if not it default to
SimplePrincipal as before ( except that SimplePrincipal is public )
        - isSecure impl moved to RequestImpl - the adapter is the only one to know if https
was used, and it should set protocol to https
	- userInRole - the auth. module will set all user groups. It is possible to make this
a callback, but only after we implement the hooks ( or events )

- moved the "tomcat-users.xml" processing to SimpleRealm - the authentication module will
only deal with authentication. Probably we can make MemoryRealm a public class, with public
API, the goal of the auth module is to implement the "adapter" pattern - between Tomcat
needs and other available auth APIs.

- DefaultCMSetter will set the right "login" servlet, based on declared auth method.

- start documenting the security implementation