Update XmlMapper to support optional validation, and optional registration of local URLs for the resolveEntity() call.  Using this feature avoids having to go out across the Internet when parsing.
Update WebXmlReader to register the local copy of the deployment
descriptor DTD file (in $TOMCAT_HOME/conf/web.dtd), and to use a
validating parser when parsing web.xml files.  This will have the side
effect of breaking Tomcat startup for applications with web.xml files that
do not conform to the DTD.

NOTE:  At this point, Tomcat's servlet container is fully compliant with
JAXP-based selection of the XML parser to be used.  The JSP implementation
still throws an exception when running the "simple custom tag example" --
I am going to look at that issue next.