- Removed HttpDate. It was supposed to be an efficient implementation for date conversion, but it wasn't. All static fields are now part of DateTool - you don't need to create SimpleDateFormat objects, just use the static ones.
The non-static part of HttpDate consisted in a Calendar used only to store the time,
plus code to parse/format it. The code was moved in MimeHeaderField ( 2 methods),
and need to be replaced with an efficient version.

- Removed code to setDate() with current time - just use setDate( System.currentTimeMillis()).
The code is not used in tomcat.

- removed toByte conversions from MimeHeader - it was using the default encoding, and
we had no control. The conversion have to be consistent with other components.

- CookieTool.FieldPosition can't be static - it's changed during parsing

- remove (unused) [] getCookies() method from Request. If you use it in your code -
there is a simple workaround or it can be added back.