Preparing for ProtocolAdapter and IO fixes.
- Removed deprecated and unused readLine method in BufferedServletInputStream
- added missing setters in Request ( every getter has now a setter, there is no
need to extend RequestImpl )
- added get/setLocalHost - it's the local host name, it may be different from
Host header. That simplifies getServerName and the protocol adapters
- added doRead(), doWrite() to Request/Response interfaces  - the methods are the
prefered way to implement in/out ( will move to ProtocolAdapter soon )
- removed printFast from ServletOS - the char-byte conversion will be fixed soon, no
need for hacks.


- Jdk12Interceptor - remove the introspection hack - if you use JDK11 just remove it,
it can't do anything. If you use JDK12 - direct call is much faster.

- TcpConnectionHandlers: add setServer( ), remove the attribute hack used to set the
server. Note that we don't use ContextManager type but object - the interface is not
tomcat specific.
- all attributes that are not interpreted by the TcpConnector ( ip, etc) are passed to
TcpConnectionHandlers. This allows better tuning and is the original purpose of setAtt().

- small fix in ajp12connectiohandler - the facade is not available till after context map.

- small change in notes - we keep a number of reserved Ids
PR:
Obtained from:
Submitted by:
Reviewed by: