Small changes to the ThreadPool - it seems the thread data mechanism didn't worked as expected.
Changed the name of the object encapsulating the action to ThreadPoolRunnable - to
avoid confusions.

What happens is the runIt() method ( or run() as it was before ) is executed and
then the thread keeps running.

This threadData mechanism is supposed to be faster than using pools - but I don't have enough
data about how much faster. Based on real numbers we can either drop this mechanism
and just use pools or enhance it to support a mix of actions ( since a running thread has
only one thread data object ).

Gal, please let me know if you find anything wrong in this patch, I can roll it back.