Initial check in for new buffer classes.
The code is factored out from existing buffers, and will replace MessageBytes,
MsgBuffer, will allow better buffering for the Input/Output streams, and will
be used in char->byte conversions.

One of the goals is to minimize the number of copy operations ( from servlet to
Writer to BufferedOutputStream to Adapter ), reduce the memory usage ( each request
will have 2 buffers for in/out and few additional buffers used by adapter ),
simplify the design ( only one type of buffer instead of 3 or 4 ).