I was offline - many changes, but most of them are just moving some code and small evolutionary changes..
Everything that worked before should still work, I just
added few more methods and hooks, preparing for 2 bigger
changes.

- added support for explicit in/out buffers using ByteBuffer.
The only change in core is adding the property to Request/Response.

This will be very important for:
  - performance - we can minimize and control the flow, probably
save 2-3 copies.
  - character encoding - this is the first step in allowing
modules to control the byte<->char and fix the wrong code
that assumes ascii ( that also means javax.servlet, but we
can override the methods ).
  - performance - enable faster character encoders
  - protocol adapters - all existing adapters implement only
the read/write operations.

- added hooks for buffer - if a buffer is present, it will be used
instead of BufferedServletInput/Output stream.

- the code in buffer is the same as in BufferedOutputStream.
BufferedInputStream was just a stub, added the buffering
( a bit better, not complete yet )

- started moving the adapters to the new model ( event callbacks
instead of extending request/response). The old model still works,
but I think we should deprecate it soon.

- Ajp13ConnectorResponse: no longer extend BufferedOS, just
override write. It's a first step in moving to ByteBuffer.

- added various utils to BuffTool - scan a byte[] directly.

- reorganized the http code to use buffers - see HttpAdapter

- small change in PoolTcpConnectionHandler - replaced
the Vector with a SimplePool ( lighter ) and also added the
option to not use a pool, just per thread data ( less sync in
critical path )

- moved "commited" from BufferedServletOS to Response, where
it should be  (since BSOS is just a facade).

- added missing setPort method in request

- small fixes in facades - if request processing fails, default
context's facade will be used. This will be important after we
fix the security and implement webapp isolation.