A number of changes:
- PoolTcpConnector - a bit of code move, but very little change. We now treat setProperty
and setAttribute the same, and we have a cleaner set of properties.
- Pass all the properties/attributes to the ConnectionHandler and SocketFactory - this
way you can configure all the connector components in the same way.
- That includes creation of a SocketFactory by passing the name of the class. Not tested,
I'll do that when we have SSLFactory ready

- added an option to validate ( by default on ) - some parsers have problems,
and we want to be able to turn it off if we want ( and it's faster without )

- fall back to old sax parser if jaxp is not available. Default is crimson, see next.

- I still can't use it with Xerces - at least if I'm not connected, it seems to
be a bug in the way it handles external entities. Crimson and the old sun parser
work fine, I'll try to find a workaround in xerces.

- in ContextManager, few debug messages ( I want to know exactly who needs tomcat home, and
minimize the need for it - for embeding tomcat ), plus set the CM in interceptors ( it
works fine if server.xml is used, but have problems if "programatic" startup is used )

- ServerSocketFactory - add a setAttribute() method, a simple way to pass info from
Connector ( like keystore ). The preffered method is still "bean-like" setters, but sometimes
it's good to have a backup.