- moved SessionSerializer to tomcat/session - removed dependencies between SessionSerializer and tomcat internals. - use ClassLoader instead of ServletLoader ( which is tomcat specific ) - added handleReload to SessionManager
The idea is to keep the Session code independent of the servlet engine implementation.
We not only want "plugable" session manager, but we also want that the session
plugin is reuseable in ( and from)  other serlvet engines. It seems NES, JRun and
probably other engines provide this feature.

SessionManager represent the requirements of tomcat ( what tomcat needs from a
session manager ). The session code should be absolutely independent of tomcat
( it shouldn't use any tomcat class). SessionManager implements the "Adapter" pattern -
it allows you to plug external session code into tomcat. It should be easy to
make adapters for other "session plugins" ( assuming they are also independent of
engine internals).

( the reverse is also true - it should be trivial to use our session managers in
other engines - by creating an adapter to their plugin interface )

( maybe someone can clean up the session code in Catalina - but it's not a big
priority, and it seem they have too many internal dependencies.
 )

- remove NoInvokerServlet - if invoker is disabled, servlet/* will be mapped
to whatever web.xml wants - it's a mistake to override this.