Fixed the JDK1.1 compiler problem, build.xml and all related files.
Right now it should compile and run fine with JDK1.1 - at least it does for me with 1.1.8

All 1.2 code is confined in modules that can be configured out.

I think the Policy-based security should still work, but I have to test it out
and verify it. I simplified the configuration and moved everything in modules,
but I need Glenn to check that and make any adjustments if needed ( now that he's
a voted commiter :-).