Various changes in policy-based security - it works again ( after 1.1 fixes ).
I tried to make the model as simple as possible - while the original code was much
more powerfull, it is probably better to keep it simple until we fully understand it.

PolicyInterceptor is the main component - it will replace SetSecurityManager
and set a SecurityManager when the engine starts ( if none is set ).
It will also add the minimal permissions to each context, and set the ProtectionDomain.

We'll use only one ProtectionDomain per web-app, and all policy is stored in tomcat.policy.
( we do a refresh before loading a new context, so you can edit it after tomcat starts
and before adding a new webapp - if you don't like the default )

All you need to do is add the PolicyInterceptor in server.xml.