Okay, I've got it - four bugs!
1) readObject/writeObject in StandardSession don't serialize
thisAccessedTime, so it ends up as 0 after deserialization - the net
result of which is the session expires rather shortly after :-(

2) My patch for SessionSerializer foolishly assumes that the request
triggering the reload has a session itself, so when the session expires
SessionSerializer ends up stuffing an invalid session in the current
request (hence the isValid failure)

3) Similar-ish to (1) readObject in StandardSession doesn't account
for the fact that the StringManger field 'sm' will start off as null -
the normal initializer doesn't run, hence the NPE when the session is
invalid. I've fixed this by setting sm in readObject - is there any
reason not to make it static?

4) I don't think this one was actually involved in the scenario Jon
described, but I also noticed that the backpointer to StandardManager in
the sessions doesn't get set after deserialization either.

The patch below fixes these issues.

--
Sean Legassick
sean@somacity.com