Few changes in error handling. - avoid Request-facade-request transformations and multiple layers. - avoid mux/demux for errors - use a different handler for each case instead of a single ErrorServlet
The new handlers are part of DefaultCMSetter, which sets the defaults for the
context. You can override them with normal servlets ( by re-defining the handler
names ), but unless the application is "trusted" it can't access internal objects.

The error processing is still a bit complex, but at least we have fewer layers and it's
easier to debug.

Also, implemented FORM based login ( finally ) - it seems to work fine ( including Pet Store!).

- added more debug messages, traces - I spend a lot of time hunting "link errors" in class
loader.