(1) Add the ability to dynamically override the Java class used to implement an ActionMapping or ActionForward when processing a particular <action> or <forward> element.  This is accomplished by adding a "className" attribute that is the fully qualified class name of the class to use for this particular entry (instead of the default, which is used if you don't specify a className attribute).  Use it like this:
	<action className="MySpecialActionMapping" .../>


(2) Improve the debugging messages logged when problems are encountered
loading resource bundles, so that we can see the actual stack traces.
Also, a check for the existence of these resources has been added to the
top of the "index.jsp" page so you will get an immediate notification that
there is some sort of problem.


(3) Separate the debugging detail level used for the Digester from the
debugging detail level used for the controller servlet itself.  Depending
on what problem you're trying to solve, you might want to turn up the
detail on one or the other, but not both.


(4) Slightly improve the documentation at the top of the Javadocs for the
Digester class.  It really needs a complete user's guide of its own, but
this is better than nothing.