Wrapper for Logger. It has a preferred log name to write to; if it can't find a log with that name, it outputs to either System.out or System.err depending on the message's warning level.  Also prepends a descriptive name to each message (usually the toString() of the calling object), so it's easier to identify the source.
Intended for use by client classes to make it easy to do reliable,
consistent logging behavior, even if you don't necessarily have a
context, if you're not in Tomcat, or if you haven't registered any log
files yet.  Not intended to supplant Logger, but to allow client
objects a consistent bit of code that prepares log messages before
they reach logger (and does the right thing if there is no logger).