Remove all the cases I can remember where two setters (one for int and one for String) are provided, per the clarification to the JSP 1.1 specs.
Why did I delete the "int" versions of the setters?  Good question -- the
answer is that keeping the "int" setters and deleting the "String" setters
does not work on Tomcat -- an issue I'm going to address separately,
because they should.

In the mean time, this fix should solve issues on Orion where the server
complained about "size" being a read-only property.  Could those who have
experienced this error please download the new version and confirm that it
has been fixed?