Few more cosmetic changes. The goal is to simplify jasper and clarify the interfaces.
- Compiler is now a normal class instead of abstract. The only abstract method is
isOutDated, and it can be overrident if needed - but the outdated check can be made outside
compiler. That will also allow to have the mangler as a separate class( simpler) .

- in JspServlet, there is no need to keep 2 hashtable keyed by JspURI, one containing
the classes and one containing a struct. We can just keep the class in the struct.

- in JspServlet, copied the isOutDated check - it will be invoked early, before the
compiler is created - thus avoid a lot of allocation if the class is not outdated.

None of the changes should have negative impact - if it does then probably it's better
to fix ( and understand what happens - I think it's better to have a bug we can fix than
code that works but we don't know why )

PR:
Obtained from:
Submitted by:
Reviewed by: