More jasper changes:
- Removed the dependency of Jasper to JspLoader - any class loader can be used
( I have a prototype using AdaptiveClassLoader, probably will not go in for
this release )

- The only "feature" that is needed by jasper is to add a jar to classpath. Instead
of creating a new interface, addJar() method is part of JspCompilationContext - where
all "context" methods are defined ( i.e. methods that have to be implemented by the
jasper container )

- JspC and JspServlet both (happen to ) use JasperLoader. Made this explicit.
Probably JspC can use any other class loader, but right now it uses JasperLoader.

I think this is my last change to jasper ( for this release ). For tomcat 3.3 we should
review the JspCompilationContext, Mangler and Options - many methods are not needed
and everything can probably be simplified a lot.

The 3 classes plus Compiler is all that's needed to interface with Jasper to translate
from jsp to java.

We should also create a separate package for jspC, make a new package for JavaCompiler
( since it's orthogonal and useable outside jasper ) and document the compiler's interfaces.