Add a new way to set optional properties on <action> and <forward> elements that is more tool friendly (i.e. you can have extended ActionForward and ActionMapping implementation classes with additional properties but still use a to-be-created DTD for the action.xml file).
The new technique lets you add as many <property> entries as you want,
nested inside an <action> element.  The previous mechanism of just adding
more attributes on the <action> element itself will continue to work for
those who prefer that approach -- but you will need to modify the
action.dtd DTD (when it is added later) for each new attribute.