Enhance the <struts:select> tag to support two new attributes:
* The "multiple" attribute causes the rendered <select> element to include
  the "multiple" modifier.  (See below for important information on this.)

* The "size" attribute lets you define how many available options are
  displayed.  If not specified, the browser default (typically 1) is used.

NOTE:  JSP custom tags must conform to the syntax requirements of XML, which
does not allow the HTML technique of "attribute minimization" (specifying an
attribute name with no value).  Thus, you must specify a value for the
"multiple" attribute, even though it is not used, like this:

	<struts:select name="serverType" multiple="multiple" size="3" ...>
	  ... available options as before ...
	</struts:select>

This enhancement is based on a suggestion by Jen Hsien Huang
<u5321007@ncnu.edu.tw>, but is implemented slightly differently than what
was proposed (I made the two attributes independent of each other).