<jsp:setProperty name="somename" property="someproperty" value=""/> was not setting the value to an empty string, it just ignored it.
The syntax for jsp:setProperty 'property' explicitely states that
if a parameter has a value of "", the corresponding property is
not set.  But the syntax for 'value' does not.  From my reading
of the specs this means that a "" value is valid as long as
the property is not being set from a parameter.

A setProperty with an empty string now works.