Checked in fix from "Warner Onstine" <onstine@intalio.com>
This is supposed to enable client authentication using certs, but I
haven't tested it yet. The patch seems ok and doesn't brake anything,
and if it does we can fix it.

If anyone can test it maybe we can merge it to 3.2, it would be usefull, but
we need more to make it work with web.xml requirements, docs, etc.

Submitted by: "Warner Onstine" <onstine@intalio.com>