This is a big change, I couldn't find a good way of spliting it ( most changes are related with each other ).
This is just a start for a big refactoring - the goal is security and
performance.

This change:

- For all internal Handlers, use a per/thread re-usable StringBuffer.
The buffer is not converted toString, so it generates no garbage.

- use OutputBuffer - the code from BufferedServletOutputStream is now here,
plus addition to write a StringBuffer ( without the expensive toString())
( still need to make sure the char/byte mixing works in all cases - this
is the DEV branch )

- removed the old ByteBuffer code - interesting experiment, too complex.

- enhanced MessageBytes ( used as an efficient way to store headers ).
Also use MessageBytes for efficient storage of request fields.
This still require a lot of work, but we are getting very close to 0 GC.
( at least for requests without params and without cookies - those will
also need to be converted to recyclable structures )

- also few small cosmetic changes in RequestImpl to allow better visibility.
Most fields will disapear soon, it will be a very simple class.

- Same for response - only OutputBuffer is used in Request and core, the
Output/Writer is handled by facade.

- few minor optimizations ( cache, avoid GC )

- a bit of cleanup ( more to come )

- few optimizations in HttpRequest