First chunk of code to simplify startup and allow customizable class loaders.
startup/Main will use URLClassLoader or SimpleClassLoader ( if jdk1.1 is
detected ) and load all the needed jars ( XXX make this customizable ).
It will also set the parentLoader to its own loader, and that can be
used ( later ) as the parent for servlets loaders ( so tomcat jars
don't interfere ).

We'll also build Main and the minimal set of required classes into tomcat.jar,
that will get a MANIFEST - in the end we should be able to replace all
the scripts with java -jar tomcat.jar ( if we find some tricks to
detect TOMCAT_HOME from java)

I also added a LoaderInterceptor11 - I think we should deprecate Adaptive
ClassLoader and all other classes - the code is way too complex ( compared
with DependManager and a simple class loader) and for 1.2 we are far
away from the security checks of URLClassLoader.

The only problem is that tomcat.policy will need to become a bit more
verbose ( that can be automated ).


startup/Main will use URLClassLoader or SimpleClassLoader ( if jdk1.1 is
detected ) and load all the needed jars ( XXX make this customizable ).
It will also set the parentLoader to its own loader, and that can be
used ( later ) as the parent for servlets loaders ( so tomcat jars
don't interfere ).

We'll also build Main and the minimal set of required classes into tomcat.jar,
that will get a MANIFEST - in the end we should be able to replace all
the scripts with java -jar tomcat.jar ( if we find some tricks to
detect TOMCAT_HOME from java)

I also added a LoaderInterceptor11 - I think we should deprecate Adaptive
ClassLoader and all other classes - the code is way too complex ( compared
with DependManager and a simple class loader) and for 1.2 we are far
away from the security checks of URLClassLoader.

The only problem is that tomcat.policy will need to become a bit more
verbose ( that can be automated ).


startup/Main will use URLClassLoader or SimpleClassLoader ( if jdk1.1 is
detected ) and load all the needed jars ( XXX make this customizable ).
It will also set the parentLoader to its own loader, and that can be
used ( later ) as the parent for servlets loaders ( so tomcat jars
don't interfere ).

We'll also build Main and the minimal set of required classes into tomcat.jar,
that will get a MANIFEST - in the end we should be able to replace all
the scripts with java -jar tomcat.jar ( if we find some tricks to
detect TOMCAT_HOME from java)

I also added a LoaderInterceptor11 - I think we should deprecate Adaptive
ClassLoader and all other classes - the code is way too complex ( compared
with DependManager and a simple class loader) and for 1.2 we are far
away from the security checks of URLClassLoader.

The only problem is that tomcat.policy will need to become a bit more
verbose ( that can be automated ).

The code consist mostly on additions, I hope it doesn't brake anything -
I need feedback and help.